рефакторинг проекта на архитектурный паттерн MVVM.
