+++ app/build.gradle
@@ -2,6 +2,9 @@ plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
     id "kotlin-parcelize"
+    id 'kotlin-kapt'
+
+
 }
 
 android {
@@ -34,6 +37,7 @@ android {
 
     buildFeatures{
         viewBinding true
+        dataBinding true
     }
 }
