+++ app/src/main/java/com/example/movi_box/view/customviews/RatingDonutView.kt
@@ -1,40 +1,30 @@
-package com.example.movi_box
+package com.example.movi_box.view.customviews
 
 import android.content.Context
 import android.graphics.*
 import android.util.AttributeSet
 import android.view.View
+import com.example.movi_box.R
 
 class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: AttributeSet? = null) : View(context, attributeSet) {
-    //Овал для рисования сегментов прогресс бара
     private val oval = RectF()
-    //Координаты центра View, а также Radius
     private var radius: Float = 0f
     private var centerX: Float = 0f
     private var centerY: Float = 0f
-    //Толщина линии прогресса
     private var stroke = 10f
-    //Значение прогресса от 0 - 100
     private var progress = 50
-    //Значения размера текста внутри кольца
     private var scaleSize = 60f
-    //Краски для наших фигур
     private lateinit var strokePaint: Paint
     private lateinit var digitPaint: Paint
     private lateinit var circlePaint: Paint
 
     private fun initPaint() {
-        //Краска для колец
         strokePaint = Paint().apply {
             style = Paint.Style.STROKE
-            //Сюда кладем значение из поля класса, потому как у нас краски будут видоизменяться
             strokeWidth = stroke
-            //Цвет мы тоже будем получать в специальном методе, потому что в зависимости от рейтинга
-            //мы будем менять цвет нашего кольца
             color = getPaintColor(progress)
             isAntiAlias = true
         }
-        //Краска для цифр
         digitPaint = Paint().apply {
             style = Paint.Style.FILL_AND_STROKE
             strokeWidth = 2f
@@ -44,7 +34,6 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
             color = getPaintColor(progress)
             isAntiAlias = true
         }
-        //Краска для заднего фона
         circlePaint = Paint().apply {
             style = Paint.Style.FILL
             color = Color.DKGRAY
@@ -59,7 +48,6 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
     }
 
     init {
-        //Получаем атрибуты и устанавливаем их в соответствующие поля
         val a =
             context.theme.obtainStyledAttributes(attributeSet, R.styleable.RatingDonutView, 0, 0)
         try {
@@ -69,7 +57,6 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
         } finally {
             a.recycle()
         }
-        //Инициализируем первоначальные краски
         initPaint()
     }
 
@@ -105,50 +92,34 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
         }
 
     private fun drawRating(canvas: Canvas) {
-        //Здесь мы можем регулировать размер нашего кольца
         val scale = radius * 0.8f
-        //Сохраняем канвас
         canvas.save()
-        //Перемещаем нулевые координаты канваса в центр, вы помните, так проще рисовать все круглое
         canvas.translate(centerX, centerY)
-        //Устанавливаем размеры под наш овал
         oval.set(0f - scale, 0f - scale, scale , scale)
-        //Рисуем задний фон(Желательно его отрисовать один раз в bitmap, так как он статичный)
         canvas.drawCircle(0f, 0f, radius, circlePaint)
-        //Рисуем "арки", из них и будет состоять наше кольцо + у нас тут специальный метод
         canvas.drawArc(oval, -90f, convertProgressToDegrees(progress), false, strokePaint)
-        //Восстанавливаем канвас
         canvas.restore()
     }
 
     private fun convertProgressToDegrees(progress: Int): Float = progress * 3.6f
 
     private fun drawText(canvas: Canvas) {
-        //Форматируем текст, чтобы мы выводили дробное число с одной цифрой после точки
         val message = String.format("%.1f", progress / 10f)
-        //Получаем ширину и высоту текста, чтобы компенсировать их при отрисовке, чтобы текст был
-        //точно в центре
         val widths = FloatArray(message.length)
         digitPaint.getTextWidths(message, widths)
         var advance = 0f
         for (width in widths) advance += width
-        //Рисуем наш текст
         canvas.drawText(message, centerX - advance / 2, centerY  + advance / 4, digitPaint)
     }
 
     override fun onDraw(canvas: Canvas) {
-        //Рисуем кольцо и задний фон
         drawRating(canvas)
-        //Рисуем цифры
         drawText(canvas)
     }
 
     fun setProgress(pr: Int) {
-        //Кладем новое значение в наше поле класса
         progress = pr
-        //Создаем краски с новыми цветами
         initPaint()
-        //вызываем перерисовку View
         invalidate()
     }
