+++ app/src/main/java/com/example/movi_box/view/fragments/WatchLaterFragment.kt
@@ -1,10 +1,11 @@
-package ru.coolhabit.firstapp
+package com.example.movi_box.view.fragments
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.movi_box.databinding.FragmentWatchLaterBinding
+import com.example.movi_box.utils.AnimationHelper
 
 
 class WatchLaterFragment : Fragment() {
