+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/abstracted"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
 
     <FrameLayout
