+++ app/src/main/res/layout/film_item.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="150dp">
     
-    <com.example.movi_box.RatingDonutView
+    <com.example.movi_box.view.customviews.RatingDonutView
         android:id="@+id/rating_donut"
         android:layout_width="50dp"
         android:layout_height="50dp"
